Supplies for School
The school year has already started and the 10B grade manager - Annie has to buy a certain number of packets of pens, packets with markers, as well as board cleaner. She is a regular client of a bookstore, so there is a discount for her, which represents some discount percentage of the total amount. Write a program that calculates how much money Annie will need to collect to pay the bill, keeping in mind the following price list:

· Package of pens - 5.80 lv.

· Package of markers - 7.20 lv.

· Board cleaner - 1.20 BGN (per liter)